// src/api/models/Watch.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Watch = {
  readonly id: number;
  readonly link: string | null;
  url: string;
  context: string | null;
  seconds: string;
  percent?: number;
  position?: number;
  readonly last_watch: string;
  user: number;
};